+++ storage/uc.go
@@ -467,26 +467,18 @@ type BucketQuota struct {
 // SetBucketQuota 设置存储空间的配额限制
 // 配额限制主要是两块， 空间存储量的限制和空间文件数限制
 func (m *BucketManager) SetBucketQuota(bucket string, size, count int64) (err error) {
-	reqHost, rErr := m.z0ApiHost()
-	if rErr != nil {
-		err = rErr
-		return
-	}
-	reqHost = strings.TrimRight(reqHost, "/")
-	reqURL := fmt.Sprintf("%s/setbucketquota/%s/size/%d/count/%d", reqHost, bucket, size, count)
+	host := getUcHost(m.Cfg.UseHTTPS)
+	host = strings.TrimRight(host, "/")
+	reqURL := fmt.Sprintf("%s/setbucketquota/%s/size/%d/count/%d", host, bucket, size, count)
 	err = m.Client.CredentialedCall(context.Background(), m.Mac, auth.TokenQiniu, nil, "POST", reqURL, nil)
 	return
 }
 
 // GetBucketQuota 获取存储空间的配额信息
 func (m *BucketManager) GetBucketQuota(bucket string) (quota BucketQuota, err error) {
-	reqHost, rErr := m.z0ApiHost()
-	if rErr != nil {
-		err = rErr
-		return
-	}
-	reqHost = strings.TrimRight(reqHost, "/")
-	reqURL := reqHost + "/getbucketquota/" + bucket
+	host := getUcHost(m.Cfg.UseHTTPS)
+	host = strings.TrimRight(host, "/")
+	reqURL := host + "/getbucketquota/" + bucket
 	err = m.Client.CredentialedCall(context.Background(), m.Mac, auth.TokenQiniu, &quota, "POST", reqURL, nil)
 	return
 }